対応再開に伴いモジュールを細分化
